ai-resume-analyzer/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
spacy==3.7.4
pydantic==2.7.1
motor==3.4.0
pymongo==4.7.2
python-dotenv==1.0.1